Allow createElement to be required without hyperx
Closes #20. This is the same thing as #20, opened against the latest
code.
