+++ README.md
@@ -46,6 +46,19 @@ module.exports = function (items) {
 }
 ```
 
+Or create an element without using template literals:
+
+```js
+// list.js
+var bel = require('bel/create-element')
+
+module.exports = function (items) {
+  return bel('ul', {}, items.map(function(item) {
+    return bel('li', {}, [item])
+  }))
+}
+```
+
 Then pass data to it and add to the DOM:
 
 ```js
